Update 2021-03-08-TheMobilizationofACTUP.md
Dear Paul-Louis,

You've done a really excellent job here. I made just a couple of very minor edits for typos (removed an extra "the" and an extra "in"), but otherwise this looks very good and is ready to go live. Your choice of subject matter lends itself nicely to a StoryMap, given that we can see these clusters of protests happening in close proximity to one another. Your Markdown syntax and StoryMap are produced well, and the research that went into this project is evident. Would you object to me posting it to our course site later today? Grade: A

A couple of other comments: as you move forward toward the final project at the end of the semester, I'd like you to think about how you can use archival materials to begin to tell not just a story, but to make an argument. That wasn't at all necessary for this assignment, but at the end of the semester I'd like to see you diving in to sources that perhaps aren't as well known, telling a story to make an argument about activism or public space or both. I think, if you're interested, you might benefit from working with Amna and José Pablo on activism around Princeton. I imagine there must be materials within the Princeton archives related to queer activism, and you could build on the subject matter of this class to perhaps even talk about how, as media technology has progressed, student activists have exploited these media to great effect. You could even reflect on the difficulty of archiving these ephemeral media sources--like Tweets, for example. These are just ideas, and I'm not trying to prescribe your final project for you, I'm just starting to get a better sense of your interests and I'm hoping to suggest ways that you can do some original research in Princeton's collections to make an argument that builds on the themes of our class and connects to a really important topic, queer activism.